+++ kedro_mlflow/framework/hooks/__init__.py
@@ -1,1 +1,2 @@
 from .mlflow_hook import MlflowHook, mlflow_hook
+from .mlflow_partitioned_hook import MlflowPartitionedHook, mlflow_partitioned_hook