+++ kedro_mlflow/utils.py
@@ -1,5 +1,9 @@
+import importlib.metadata
+from functools import lru_cache
 from pathlib import Path
-from typing import List, Union
+from typing import Any, Callable, Dict, List, Union
+
+import flatten_dict
 
 
 def _parse_requirements(path: Union[str, Path], encoding="utf-8") -> List:
@@ -8,3 +12,25 @@ def _parse_requirements(path: Union[str, Path], encoding="utf-8") -> List:
             x.strip() for x in file_handler if x.strip() and not x.startswith("-r")
         ]
     return requirements
+
+
+@lru_cache(maxsize=None)
+def _load_plugins(entry_point: str) -> Dict[str, Callable[[], Any]]:
+    return {
+        plugin.name: plugin.load
+        for plugin in importlib.metadata.entry_points()[entry_point]
+    }
+def _flatten_dict(d: Dict, recursive: bool = True, sep: str = ".") -> Dict:
+    def reducer(k1: str, k2: str):
+        return f"{k1}{sep}{k2}" if k1 else str(k2)
+
+    return flatten_dict.flatten(
+        d, reducer=reducer, max_flatten_depth=(None if recursive else 2)
+    )
+
+
+def _unflatten_dict(d: Dict, sep: str = ".") -> Dict:
+    def splitter(k: str):
+        return k.split(sep)
+
+    return flatten_dict.unflatten(d, splitter=splitter)