+++ tests/framework/hooks/test_hook_deactivate_tracking.py
@@ -281,9 +281,9 @@ def test_deactivated_tracking_but_not_for_given_pipeline(
         # 0 is default, 1 is "fake_exp"
         all_runs_id_beginning = set(
             [
-                run.run_id
-                for k in range(len(mlflow_client.list_experiments()))
-                for run in mlflow_client.list_run_infos(experiment_id=f"{k}")
+                run.info.run_id
+                for k in range(len(mlflow_client.search_experiments()))
+                for run in mlflow_client.search_runs(experiment_ids=[f"{k}"])
             ]
         )
 
@@ -291,9 +291,9 @@ def test_deactivated_tracking_but_not_for_given_pipeline(
 
         all_runs_id_end = set(
             [
-                run.run_id
-                for k in range(len(mlflow_client.list_experiments()))
-                for run in mlflow_client.list_run_infos(experiment_id=f"{k}")
+                run.info.run_id
+                for k in range(len(mlflow_client.search_experiments()))
+                for run in mlflow_client.search_runs(experiment_ids=[f"{k}"])
             ]
         )
 
@@ -312,9 +312,9 @@ def test_deactivated_tracking_for_given_pipeline(
         # 0 is default, 1 is "fake_exp"
         all_runs_id_beginning = set(
             [
-                run.run_id
-                for k in range(len(mlflow_client.list_experiments()))
-                for run in mlflow_client.list_run_infos(experiment_id=f"{k}")
+                run.info.run_id
+                for k in range(len(mlflow_client.search_experiments()))
+                for run in mlflow_client.search_runs(experiment_ids=[f"{k}"])
             ]
         )
 
@@ -322,9 +322,9 @@ def test_deactivated_tracking_for_given_pipeline(
 
         all_runs_id_end = set(
             [
-                run.run_id
-                for k in range(len(mlflow_client.list_experiments()))
-                for run in mlflow_client.list_run_infos(experiment_id=f"{k}")
+                run.info.run_id
+                for k in range(len(mlflow_client.search_experiments()))
+                for run in mlflow_client.search_runs(experiment_ids=[f"{k}"])
             ]
         )